% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reading.R
\name{read_hyperfine_json}
\alias{read_hyperfine_json}
\title{Read hyperfine timing results in JSON format}
\usage{
read_hyperfine_json(filename)
}
\arguments{
\item{filename}{: a JSON file in the format written by hyperfine}
}
\value{
a tibble with two columns
}
\description{
Read hyperfine timing results in JSON format
}
